requirements.txt
click==7.1.2
Flask==0.10.1
Flask-API.yandex==0.6.2.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==0.16.1
